requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
qrcode>=7.4.0
Pillow>=10.0.0
streamlit-option-menu>=0.3.6
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0